+++ src/Heartbeat/Sensor/DBConnection.php
@@ -35,4 +35,12 @@ class DBConnection extends Sensor
             return false;
         }
     }
+
+    /**
+     * @inheritDoc
+     */
+    protected function getStatusMessage(bool $status): string
+    {
+        return $status ? __d('Heartbeat', 'OK') : __d('Heartbeat', 'FAILED');
+    }
 }