+++ src/Heartbeat/Sensor/DebugMode.php
@@ -18,4 +18,12 @@ class DebugMode extends Sensor
     {
         return (bool)Configure::read('debug');
     }
+
+    /**
+     * @inheritDoc
+     */
+    protected function getStatusMessage(bool $status): string
+    {
+        return $status ? __d('Heartbeat', 'OK') : __d('Heartbeat', 'FAILED');
+    }
 }