+++ tests/TestCase/Sensor/DummySensor.php
@@ -17,4 +17,12 @@ class DummySensor extends Sensor
     {
         return true;
     }
+
+    /**
+     * @inheritDoc
+     */
+    protected function getStatusMessage(bool $status): string
+    {
+        return $status ? __d('Heartbeat', 'OK') : __d('Heartbeat', 'FAILED');
+    }
 }